+++ src/pyneapple/results/results.py
@@ -17,172 +17,31 @@ import numpy as np
 from pathlib import Path
 import pandas as pd
 import matplotlib.pyplot as plt
+
 # from nifti import NiiFit, Nii
 from radimgarray import RadImgArray, SegImgArray
+from . import ResultDict
 
 if TYPE_CHECKING:
     from . import FitData
-
-
-class CustomDict(dict):
-    """Custom dictionary for storing fitting results and returning them according to
-    fit style.
-
-    Basic dictionary enhanced with fit type utility to store results of pixel or
-    segmentation wise fitting. Some methods are overwritten to handle either of the
-    two fit types.
-
-    Attributes:
-        fit_type (str): Type of fit process. Either "Pixel" or "Segmentation".
-        identifier (dict): Dictionary containing pixel to segmentation value pairs.
-
-    Methods:
-        set_segmentation_wise(self, identifier: dict) Update the dictionary for
-            segmented fitting
-        as_array(self, shape: dict) -> np.ndarray Return array containing the dict
-            content
-    """
-
-    def __init__(self, fit_type: str | None = None, identifier: dict | None = None):
-        """Initialize CustomDict object.
-
-        Args:
-            fit_type (str): Type of fit process. Either "Pixel" or "Segmentation".
-            identifier (dict): Dictionary containing pixel to segmentation value pairs.
-        """
-        super().__init__()
-        self.type = fit_type
-        self.identifier = identifier
-        if fit_type == "Segmentation" and identifier is None:
-            raise ValueError("Identifier is required if fit_type is 'Segmentation'")
-
-    def __getitem__(self, key):
-        """Return value of key in dictionary.
-
-        The __getitem__ method is overwritten to handle either pixel or segmentation
-        based keys.
-
-        Args:
-            key: Key to look up in dictionary.
-        Returns:
-            value: Value of key in dictionary.
-        """
-        value = None
-        key = self.validate_key(key)
-        if isinstance(key, tuple):
-            # If the key is a tuple containing the pixel coordinates:
-            try:
-                if self.type == "Segmentation":
-                    # in case of Segmentation wise fitting the identifier
-                    # dict is needed to look up pixel segmentation number
-                    value = super().__getitem__(self.identifier[key])
-                else:
-                    value = super().__getitem__(key)
-            except KeyError:
-                raise KeyError(f"Key '{key}' not found in dictionary.")
-        elif isinstance(key, int):
-            # If the key is an int for the segmentation:
-            try:
-                if self.type == "Segmentation":
-                    value = super().__getitem__(key)
-            except KeyError:
-                raise KeyError(f"Key '{key}' not found in dictionary.")
-        return value
-
-    def __setitem__(self, key, value):
-        """Set value of key in dictionary."""
-        key = self.validate_key(key)
-        super().__setitem__(key, value)
-
-    def get(self, key, default=None):
-        """Return value of key in dictionary."""
-        try:
-            return self.__getitem__(key)
-        except KeyError:
-            if default is not None:
-                return default
-            else:
-                KeyError(f"Key '{key}' not found in dictionary.")
-
-    @staticmethod
-    def validate_key(key):
-        """Validate key type.
-        Check weather the given key is supported by the CustomDict."""
-        if isinstance(key, tuple):
-            pass
-        elif isinstance(key, int):
-            pass
-        elif isinstance(key, str):
-            TypeError("String assignment and calling is not supported.")
-        elif isinstance(key, float):
-            TypeError("Float assignment and calling is not supported.")
-        try:
-            if np.issubdtype(key, np.integer):
-                key = int(key)
-        except TypeError:
-            pass
-        return key
-
-    def set_segmentation_wise(self, identifier: dict | None = None):
-        """
-        Update segmentation info of dict.
-
-        Args:
-            identifier (dict): Dictionary containing pixel to segmentation value pairs.
-        """
-        if isinstance(identifier, dict):
-            self.identifier = identifier  # .copy()
-            self.type = "Segmentation"
-        elif identifier is None or False:
-            self.identifier = {}
-            self.type = "Pixel"
-
-    def as_array(self, shape: tuple | list) -> np.ndarray:
-        """Returns a numpy array of the dict fit data.
-
-        Args:
-            shape (tuple): Shape of final fit data.
-
-        Returns:
-            array (np.ndarray): Numpy array of the dict fit data.
-        """
-        if isinstance(shape, tuple):
-            shape = list(shape)
-
-        if len(shape) < 4:
-            ValueError("Shape must be at least 4 dimensions.")
-        elif shape[3] == 1:
-            shape[3] = list(self.values())[0].shape[
-                0
-            ]  # read shape of first array in dict to determine shape
-            pass
-        array = np.zeros(shape)
-
-        if self.type == "Segmentation":
-            for key, seg_number in self.identifier.items():
-                array[key] = self[seg_number]
-        else:
-            for key, value in self.items():
-                array[key] = value
-        return array
 
 
 class Results:
     """Class containing estimated diffusion values and fractions.
 
     Attributes:
-        d (CustomDict): Dict of tuples containing pixel coordinates as keys and a
+        d (ResultDict): Dict of tuples containing pixel coordinates as keys and a
             np.ndarray holding all the d values.
-        f (CustomDict): Dict of tuples containing pixel coordinates as keys and a
+        f (ResultDict): Dict of tuples containing pixel coordinates as keys and a
             np.ndarray holding all the f values.
-        S0 (CustomDict): Dict of tuples containing pixel coordinates as keys and a
+        s_0 (ResultDict): Dict of tuples containing pixel coordinates as keys and a
             np.ndarray holding all the S0 values.
-        spectrum (CustomDict): Dict of tuples containing pixel coordinates as keys and a
+        spectrum (ResultDict): Dict of tuples containing pixel coordinates as keys and a
             np.ndarray holding all the spectrum values.
-        curve (CustomDict): Dict of tuples containing pixel coordinates as keys and a
+        curve (ResultDict): Dict of tuples containing pixel coordinates as keys and a
             np.ndarray holding all the curve values.
-        raw (CustomDict): Dict holding raw fit data.
-        T1 (CustomDict): Dict of tuples containing pixel coordinates as keys and a
+        raw (ResultDict): Dict holding raw fit data.
+        t_1 (ResultDict): Dict of tuples containing pixel coordinates as keys and a
             np.ndarray holding all the T1 values.
 
     Methods:
@@ -201,15 +60,15 @@ class Results:
             PNG files. If no slice_number is passed, plots the first slice.
     """
 
-    def __init__(self):
+    def __init__(self, **kwargs):
         """Initialize Results object."""
-        self.spectrum: CustomDict = CustomDict()
-        self.curve: CustomDict = CustomDict()
-        self.raw: CustomDict = CustomDict()
-        self.d: CustomDict = CustomDict()
-        self.f: CustomDict = CustomDict()
-        self.S0: CustomDict = CustomDict()
-        self.T1: CustomDict = CustomDict()
+        self.spectrum: ResultDict = ResultDict()
+        self.curve: ResultDict = ResultDict()
+        self.raw: ResultDict = ResultDict()  # is this actually a thing anymore?
+        self.d: ResultDict = ResultDict()
+        self.f: ResultDict = ResultDict()
+        self.s_0: ResultDict = ResultDict()
+        self.t_1: ResultDict = ResultDict()
 
     def set_segmentation_wise(self, identifier: dict):
         """Set segmentation info of all dicts.
@@ -385,7 +244,7 @@ class Results:
         for key in self.d:
             array[key[0], key[1], key[2], 0:n_components] = self.d[key]
             array[key[0], key[1], key[2], n_components:-1] = self.f[key]
-            array[key[0], key[1], key[2], -1] = self.S0[key]
+            array[key[0], key[1], key[2], -1] = self.s_0[key]
 
         # TODO: needs rework
         # out_nii = NiiFit(n_components=n_components).from_array(array)
